refactor: extract helperSort
